fix: echo the task's context
Since the context tool isn't using sys.echo, the prompt is fed to the
LLM and the response of is being used as the context. This is not the
intended behavior.

Using sys.echo here will mean that the text is used in the system prompt
and the LLM is not used.

Signed-off-by: Donnie Adams <donnie@acorn.io>