fix: fall back to utc for invalid timezones
In some situations the `Current Date and Time` context tool
is called without `OBOT_USER_TIMEZONE` environment being set by Obot,
causing the call to fail with an invalid timezone error.
In particular, this happens when configuring agent-scoped auth for tools
that share `Current Date and Time`, and breaks the feature.

To fix this, parse the timezone in the `Current Date and Time` and
`User Timezone` tools and fall back to UTC when it's invalid.

Addresses https://github.com/obot-platform/obot/issues/1157

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>