enhance: use credentials to capture sendgrid email address
Email addresses and API keys are tightly coupled for the `SendGrid` API;
i.e. you can only send emails from authorized addresses associated with
the key you're using.

To codify this relationship in the `SendGrid` tool, share an additional credential
to capture the email address alongside the API key.

Additionally, add an optional `from_name` parameter to the `Send Email` tool
to support for setting the sender name. Default's to `Obot` if not set.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>