feat: add support to use PATs instead of OAuth
Tools will now be able to specify, in their metadata, that their OAuth
credential supports personal access tokens with the oauthPATSupported
field being "true". If a tool does this, then the OAuth credential will
prompt the user for a PAT instead of going through the OAuth flow.

This change includes this field on the GitHub and Atlassian tools.

Lastly, this change removes support for OAuth outside Obot. The OAuth
credential tool will no longer look at the CLI config nor default to use
the Gateway.

Signed-off-by: Donnie Adams <donnie@acorn.io>