enhance: add validate tool for github auth provider
Add a `validate` tool that can be used to validate options passed to the
github-auth-provider.

The tool validates each option field and returns a JSON object
containing any validation errors it encountered. Each error in the response
is keyed to the associated field name with the goal of driving better
form UX in Obot's UI.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>