chore: use OBOT_SERVER_PUBLIC_URL where necessary
The OBOT_SERVER_URL will be a localhost URL and is intended for tools that are
fork/exec-ed in the Obot process to talk back to Obot. This doesn't work for
things like auth providers because they need to redirect the user's browser to
the public URL.

This change uses the public URL where necessary.